Create nodes for a step-package Sankey diagram

Create nodes for a step-package Sankey diagram

Usage

make_nodes(x)

Arguments

x

A data frame containing at least a package column where each entry describes the package combination at a given step.

Value

A data frame with columns name, node_name, and id suitable for use with networkD3::sankeyNetwork().

Examples

df <- simulate(3)
make_nodes(df)
#>                    name             node_name tooltip package_id  colour
#> 1                    cm                    cm    100%          1 #AEC6CF
#> 2               cm, itn               cm\nitn    100%          8 #FFB347
#> 3          cm, itn, irs          cm\nitn\nirs   33.3%         11 #FF6961
#> 4          cm, itn, smc          cm\nitn\nsmc   66.7%       <NA>    <NA>
#> 5     cm, itn, smc, irs     cm\nitn\nsmc\nirs   66.7%       <NA>    <NA>
#> 6      cm, itn, smc, vx      cm\nitn\nsmc\nvx   33.3%       <NA>    <NA>
#> 7 cm, itn, smc, vx, irs cm\nitn\nsmc\nvx\nirs    100%       <NA>    <NA>
#> 8                 ghost                  <NA>   ghost       <NA>    <NA>